Reset shadcn component
